feat(admin, oauth): consolidate OAuth client management under unified interface
Moved and refactored OAuth client
